+++ HISTORY.rst
@@ -9,8 +9,8 @@ picows Release History
 ------------------
 
 * ws_connect/ws_create_server logger_name parameter can now accept a logger-like object
+* ws_connect/ws_create_server websocket_handshake_timeout param can now accept None to disable handshake timeouts
 * Introduce new exceptions: WSInvalidMessageError, WSInvalidStatusError, WSInvalidHeaderError, WSInvalidUpgradeError
-
 
 1.19.0 (2026-04-24)
 ------------------